+++ client/src/app/account/page.tsx
@@ -2,12 +2,13 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import Image from "next/image";
 import SignInForm from "@/components/sign-in-form";
 import SignUpForm from "@/components/sign-up-form";
+import Logo from "@/assets/logo-1.png";
 
 export default function RegisterPage() {
   return (
     <>
       <div className="flex min-h-screen flex-col justify-center items-center bg-gradient-to-b from-[#eefeff] to-white gap-4">
-        <Image src={"/logo-1.png"} alt="image-logo" width={100} height={100} />
+        <Image src={Logo} alt="image-logo" width={100} height={100} />
 
         <p className="bg-amber-300 rounded-md px-2 py-1 text-sm font-semibold">
           Welcome to your journey
